Final Version Upload
Using Ms. Healey's review of my draft game, I fixed many aspects of my game/code that was lacking. For example, I limited the size of the board and # of bombs; I moved the "main code" to the bottom of the code; I compartmentalized the "user guess" function; added a play again feature; and allowed the user to guess without entering "g".